fix show more
