+++ index.tsx
@@ -10,33 +10,29 @@ export const Native = getNative();
 
 import "./styles.css";
 
-import { addContextMenuPatch, NavContextMenuPatchCallback, removeContextMenuPatch } from "@api/ContextMenu";
-import { definePluginSettings, Settings } from "@api/Settings";
 import ErrorBoundary from "@components/ErrorBoundary";
 import { Devs } from "@utils/constants";
 import { Logger } from "@utils/Logger";
-import definePlugin, { OptionType } from "@utils/types";
+import definePlugin from "@utils/types";
 import { findByPropsLazy } from "@webpack";
-import { Alerts, Button, FluxDispatcher, Menu, MessageActions, MessageStore, React, Toasts, UserStore } from "@webpack/common";
+import { FluxDispatcher, MessageStore, React, UserStore } from "@webpack/common";
 
 import { OpenLogsButton } from "./components/LogsButton";
 import { openLogModal } from "./components/LogsModal";
-import { ImageCacheDir, LogsDir } from "./components/settings/FolderSelectInput";
-import { openUpdaterModal } from "./components/UpdaterModal";
 import * as idb from "./db";
 import { addMessage } from "./LoggedMessageManager";
 import * as LoggedMessageManager from "./LoggedMessageManager";
+import { settings } from "./settings";
 import { FetchMessagesResponse, LoadMessagePayload, LoggedMessage, LoggedMessageJSON, MessageCreatePayload, MessageDeleteBulkPayload, MessageDeletePayload, MessageUpdatePayload, PatchAttachmentItem } from "./types";
-import { addToXAndRemoveFromOpposite, cleanUpCachedMessage, cleanupUserObject, doesBlobUrlExist, getNative, isGhostPinged, ListType, mapEditHistory, messageJsonToMessageClass, reAddDeletedMessages, removeFromX } from "./utils";
-import { DEFAULT_IMAGE_CACHE_DIR } from "./utils/constants";
+import { cleanUpCachedMessage, cleanupUserObject, doesBlobUrlExist, getNative, isGhostPinged, mapEditHistory, messageJsonToMessageClass, reAddDeletedMessages } from "./utils";
+import { removeContextMenuBindings, setupContextMenuPatches } from "./utils/contextMenu";
 import { shouldIgnore } from "./utils/index";
 import { LimitedMap } from "./utils/LimitedMap";
 import { doesMatch } from "./utils/parseQuery";
 import * as imageUtils from "./utils/saveImage";
 import * as ImageManager from "./utils/saveImage/ImageManager";
-import { exportLogs, importLogs } from "./utils/settingsUtils";
 import { checkForUpdatesAndNotify } from "./utils/updater";
-
+export { settings };
 
 export const Flogger = new Logger("MessageLoggerEnhanced", "#f26c6c");
 
@@ -193,9 +189,9 @@ async function processMessageFetch(response: FetchMessagesResponse) {
         }
 
         const firstMessage = response.body[response.body.length - 1];
-        console.time("fetching messages from idb");
+        // console.time("fetching messages from idb");
         const messages = await idb.getMessagesByChannelAndAfterTimestampIDB(firstMessage.channel_id, firstMessage.timestamp);
-        console.timeEnd("fetching messages from idb");
+        // console.timeEnd("fetching messages from idb");
 
         if (!messages.length) return;
 
@@ -236,226 +232,6 @@ async function processMessageFetch(response: FetchMessagesResponse) {
         Flogger.error("Failed to fetch messages", e);
     }
 }
-
-export const settings = definePluginSettings({
-    checkForUpdate: {
-        type: OptionType.COMPONENT,
-        description: "Check for update",
-        component: () =>
-            <Button onClick={() => openUpdaterModal()}>
-                Check For Updates
-            </Button>
-    },
-    saveMessages: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Wether to save the deleted and edited messages.",
-    },
-
-    saveImages: {
-        type: OptionType.BOOLEAN,
-        description: "Save deleted messages",
-        default: false
-    },
-
-    sortNewest: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Sort logs by newest.",
-    },
-
-    cacheMessagesFromServers: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "Usually message logger only logs from whitelisted ids and dms, enabling this would mean it would log messages from all servers as well. Note that this may cause the cache to exceed its limit, resulting in some messages being missed. If you are in a lot of servers, this may significantly increase the chances of messages being logged, which can result in a large message record and the inclusion of irrelevant messages.",
-    },
-
-    autoCheckForUpdates: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Automatically check for updates on startup.",
-    },
-
-    ignoreBots: {
-        type: OptionType.BOOLEAN,
-        description: "Whether to ignore messages by bots",
-        default: false,
-        onChange() {
-            // we will be handling the ignoreBots now (enabled or not) so the original messageLogger shouldnt
-            Settings.plugins.MessageLogger.ignoreBots = false;
-        }
-    },
-
-    ignoreSelf: {
-        type: OptionType.BOOLEAN,
-        description: "Whether to ignore messages by yourself",
-        default: false,
-        onChange() {
-            Settings.plugins.MessageLogger.ignoreSelf = false;
-        }
-    },
-
-    ignoreMutedGuilds: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "Messages in muted guilds will not be logged. Whitelisted users/channels in muted guilds will still be logged."
-    },
-
-    ignoreMutedCategories: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "Messages in channels belonging to muted categories will not be logged. Whitelisted users/channels in muted guilds will still be logged."
-    },
-
-    ignoreMutedChannels: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "Messages in muted channels will not be logged. Whitelisted users/channels in muted guilds will still be logged."
-    },
-
-    alwaysLogDirectMessages: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Always log DMs",
-    },
-
-    alwaysLogCurrentChannel: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Always log current selected channel. Blacklisted channels/users will still be ignored.",
-    },
-
-    permanentlyRemoveLogByDefault: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "Vencord's base MessageLogger remove log button wiil delete logs permanently",
-    },
-
-    hideMessageFromMessageLoggers: {
-        default: false,
-        type: OptionType.BOOLEAN,
-        description: "When enabled, a context menu button will be added to messages to allow you to delete messages without them being logged by other loggers. Might not be safe, use at your own risk."
-    },
-
-    ShowLogsButton: {
-        default: true,
-        type: OptionType.BOOLEAN,
-        description: "Toggle to whenever show the toolbox or not",
-        restartNeeded: true,
-    },
-
-    hideMessageFromMessageLoggersDeletedMessage: {
-        default: "redacted eh",
-        type: OptionType.STRING,
-        description: "The message content to replace the message with when using the hide message from message loggers feature.",
-    },
-
-    messageLimit: {
-        default: 200,
-        type: OptionType.NUMBER,
-        description: "Maximum number of messages to save. Older messages are deleted when the limit is reached. 0 means there is no limit"
-    },
-
-    imagesLimit: {
-        default: 100,
-        type: OptionType.NUMBER,
-        description: "Maximum number of images to save. Older images are deleted when the limit is reached. 0 means there is no limit"
-    },
-
-    cacheLimit: {
-        default: 1000,
-        type: OptionType.NUMBER,
-        description: "Maximum number of messages to store in the cache. Older messages are deleted when the limit is reached. This helps reduce memory usage and improve performance. 0 means there is no limit",
-    },
-
-    whitelistedIds: {
-        default: "",
-        type: OptionType.STRING,
-        description: "Whitelisted server, channel, or user IDs."
-    },
-
-    blacklistedIds: {
-        default: "",
-        type: OptionType.STRING,
-        description: "Blacklisted server, channel, or user IDs."
-    },
-
-    imageCacheDir: {
-        type: OptionType.COMPONENT,
-        description: "Select saved images directory",
-        component: ErrorBoundary.wrap(ImageCacheDir) as any
-    },
-
-    logsDir: {
-        type: OptionType.COMPONENT,
-        description: "Select logs directory",
-        component: ErrorBoundary.wrap(LogsDir) as any
-    },
-
-    importLogs: {
-        type: OptionType.COMPONENT,
-        description: "Import Logs From File",
-        component: () =>
-            <Button onClick={importLogs}>
-                Import Logs
-            </Button>
-    },
-
-    exportLogs: {
-        type: OptionType.COMPONENT,
-        description: "Export Logs From IndexedDB",
-        component: () =>
-            <Button onClick={exportLogs}>
-                Export Logs
-            </Button>
-    },
-
-    openLogs: {
-        type: OptionType.COMPONENT,
-        description: "Open Logs",
-        component: () =>
-            <Button onClick={() => openLogModal()}>
-                Open Logs
-            </Button>
-    },
-    openImageCacheFolder: {
-        type: OptionType.COMPONENT,
-        description: "Opens the image cache directory",
-        component: () =>
-            <Button
-                disabled={
-                    IS_WEB
-                    || settings.store.imageCacheDir == null
-                    || settings.store.imageCacheDir === DEFAULT_IMAGE_CACHE_DIR
-                }
-                onClick={() => Native.showItemInFolder(settings.store.imageCacheDir)}
-            >
-                Open Image Cache Folder
-            </Button>
-    },
-
-    clearLogs: {
-        type: OptionType.COMPONENT,
-        description: "Clear Logs",
-        component: () =>
-            <Button
-                color={Button.Colors.RED}
-                onClick={() => Alerts.show({
-                    title: "Clear Logs",
-                    body: "Are you sure you want to clear all logs?",
-                    confirmColor: Button.Colors.RED,
-                    confirmText: "Clear",
-                    cancelText: "Cancel",
-                    onConfirm: () => {
-                        idb.clearMessagesIDB();
-                    },
-                })}
-            >
-                Clear Logs
-            </Button>
-    },
-
-});
 
 export default definePlugin({
     name: "MessageLoggerEnhanced",
@@ -603,12 +379,10 @@ export default definePlugin({
                 return;
             }
 
-
             if (!this.attachments.has(attachment.id)) {
                 blobUrl += "#";
                 this.attachments.set(attachment.id, blobUrl);
             }
-
 
             props.item.downloadUrl = blobUrl;
             props.item.originalItem.url = blobUrl;
@@ -646,10 +420,8 @@ export default definePlugin({
     },
 
     async start() {
-        // if (!settings.store.saveMessages)
-        //     clearLogs();
-
         this.oldGetMessage = oldGetMessage = MessageStore.getMessage;
+
         // we have to do this because the original message logger fetches the message from the store now
         MessageStore.getMessage = (channelId: string, messageId: string) => {
             const MLMessage = idb.cachedMessages.get(messageId);
@@ -666,166 +438,12 @@ export default definePlugin({
         settings.store.imageCacheDir = imageCacheDir;
         settings.store.logsDir = logsDir;
 
-        addContextMenuPatch("message", contextMenuPath);
-        addContextMenuPatch("channel-context", contextMenuPath);
-        addContextMenuPatch("user-context", contextMenuPath);
-        addContextMenuPatch("guild-context", contextMenuPath);
-        addContextMenuPatch("gdm-context", contextMenuPath);
+        setupContextMenuPatches();
     },
 
     stop() {
-        // if (!settings.store.saveMessages)
-        //     clearLogs();
-
-        removeContextMenuPatch("message", contextMenuPath);
-        removeContextMenuPatch("channel-context", contextMenuPath);
-        removeContextMenuPatch("user-context", contextMenuPath);
-        removeContextMenuPatch("guild-context", contextMenuPath);
-        removeContextMenuPatch("gdm-context", contextMenuPath);
-
+        removeContextMenuBindings();
         MessageStore.getMessage = this.oldGetMessage;
     }
 });
 
-
-const idFunctions = {
-    Server: props => props?.guild?.id,
-    User: props => props?.message?.author?.id || props?.user?.id,
-    Channel: props => props.message?.channel_id || props.channel?.id
-} as const;
-
-type idKeys = keyof typeof idFunctions;
-
-function renderListOption(listType: ListType, IdType: idKeys, props: any) {
-    const id = idFunctions[IdType](props);
-    if (!id) return null;
-
-    const isBlocked = settings.store[listType].includes(id);
-    const oppositeListType = listType === "blacklistedIds" ? "whitelistedIds" : "blacklistedIds";
-    const isOppositeBlocked = settings.store[oppositeListType].includes(id);
-    const list = listType === "blacklistedIds" ? "Blacklist" : "Whitelist";
-
-    const addToList = () => addToXAndRemoveFromOpposite(listType, id);
-    const removeFromList = () => removeFromX(listType, id);
-
-    return (
-        <Menu.MenuItem
-            id={`${listType}-${IdType}-${id}`}
-            label={
-                isOppositeBlocked
-                    ? `Move ${IdType} to ${list}`
-                    : isBlocked ? `Remove ${IdType} From ${list}` : `${list} ${IdType}`
-            }
-            action={isBlocked ? removeFromList : addToList}
-        />
-    );
-}
-
-function renderOpenLogs(idType: idKeys, props: any) {
-    const id = idFunctions[idType](props);
-    if (!id) return null;
-
-    return (
-        <Menu.MenuItem
-            id={`open-logs-for-${idType.toLowerCase()}`}
-            label={`Open Logs For ${idType}`}
-            action={() => openLogModal(`${idType.toLowerCase()}:${id}`)}
-        />
-    );
-}
-
-const contextMenuPath: NavContextMenuPatchCallback = (children, props) => {
-    if (!props) return;
-
-    if (!children.some(child => child?.props?.id === "message-logger")) {
-        children.push(
-            <Menu.MenuSeparator />,
-            <Menu.MenuItem
-                id="message-logger"
-                label="Message Logger"
-            >
-
-                <Menu.MenuItem
-                    id="open-logs"
-                    label="Open Logs"
-                    action={() => openLogModal()}
-                />
-
-                {Object.keys(idFunctions).map(IdType => renderOpenLogs(IdType as idKeys, props))}
-
-                <Menu.MenuSeparator />
-
-                {Object.keys(idFunctions).map(IdType => (
-                    <React.Fragment key={IdType}>
-                        {renderListOption("blacklistedIds", IdType as idKeys, props)}
-                        {renderListOption("whitelistedIds", IdType as idKeys, props)}
-                    </React.Fragment>
-                ))}
-
-                {
-                    props.navId === "message"
-                    && (props.message?.deleted || props.message?.editHistory?.length > 0)
-                    && (
-                        <>
-                            <Menu.MenuSeparator />
-                            <Menu.MenuItem
-                                id="remove-message"
-                                label={props.message?.deleted ? "Remove Message (Permanent)" : "Remove Message History (Permanent)"}
-                                color="danger"
-                                action={() =>
-                                    idb.deleteMessageIDB(props.message.id)
-                                        .then(() => {
-                                            if (props.message.deleted) {
-                                                FluxDispatcher.dispatch({
-                                                    type: "MESSAGE_DELETE",
-                                                    channelId: props.message.channel_id,
-                                                    id: props.message.id,
-                                                    mlDeleted: true
-                                                });
-                                            } else {
-                                                props.message.editHistory = [];
-                                            }
-                                        }).catch(() => Toasts.show({
-                                            type: Toasts.Type.FAILURE,
-                                            message: "Failed to remove message",
-                                            id: Toasts.genId()
-                                        }))
-
-                                }
-                            />
-                        </>
-                    )
-                }
-
-                {
-                    settings.store.hideMessageFromMessageLoggers
-                    && props.navId === "message"
-                    && props.message?.author?.id === UserStore.getCurrentUser().id
-                    && props.message?.deleted === false
-                    && (
-                        <>
-                            <Menu.MenuSeparator />
-                            <Menu.MenuItem
-                                id="hide-from-message-loggers"
-                                label="Delete Message (Hide From Message Loggers)"
-                                color="danger"
-
-                                action={async () => {
-                                    await MessageActions.deleteMessage(props.message.channel_id, props.message.id);
-                                    MessageActions._sendMessage(props.message.channel_id, {
-                                        "content": settings.store.hideMessageFromMessageLoggersDeletedMessage,
-                                        "tts": false,
-                                        "invalidEmojis": [],
-                                        "validNonShortcutEmojis": []
-                                    }, { nonce: props.message.id });
-                                }}
-
-                            />
-                        </>
-                    )
-                }
-
-            </Menu.MenuItem>
-        );
-    }
-};
